requirements.txt
pyyaml
 --extra-index-url https://erc.timetoact-group.at/
erc3>=1.1.0
openai>=2.8.1